+++ vortex/utils/data/loader/__init__.py
@@ -3,14 +3,14 @@ import inspect
 from typing import Union
 from types import ModuleType
 
-supported_transforms = {}
-ALL_TRANSFORMS = []
+supported_loaders = {}
+wrapper_format = {}
+ALL_LOADERS = []
 
 _REQUIRED_ATTRIBUTES = [
-    'create_transform',
-    'supported_transforms'
+    'create_loader',
+    'supported_loaders'
 ]
-
 
 def register_module(module: Union[str,ModuleType]):
     """
@@ -18,7 +18,7 @@ def register_module(module: Union[str,ModuleType]):
     Args:
         module: str or module-type to be registered
     """
-    global supported_transforms, ALL_TRANSFORMS
+    global supported_loaders, ALL_LOADERS
     # TODO : consider to check module existence before importing
     if isinstance(module, str):
         # when string is passed, assume it's relative
@@ -32,29 +32,32 @@ def register_module(module: Union[str,ModuleType]):
         if not attribute in module_attributes:
             raise RuntimeError("dear maintainer, your module(s) is supposed to have the following attribute(s) : %s; but %s is missing" % (
                 _REQUIRED_ATTRIBUTES, attribute))
-    supported_transforms[module] = module.supported_transforms
-    ALL_TRANSFORMS.extend(module.supported_transforms)
+    supported_loaders[module] = [loader[0] for loader in module.supported_loaders]
+    wrapper_format.update({a:b for a,b in module.supported_loaders})
+    ALL_LOADERS.extend([loader[0] for loader in module.supported_loaders])
 
 
-def create_transform(transform: str, *args, **kwargs):
-    if not transform in ALL_TRANSFORMS:
-        raise KeyError("transfomr %s not supported, available : %s" %
-                       (transform, ALL_TRANSFORMS))
-    for module, transforms in supported_transforms.items():
-        if transform in transforms:
-            return module.create_transform(*args, **kwargs)
+def create_loader(loader: str, *args, **kwargs):
+    if not loader in ALL_LOADERS:
+        raise KeyError("dataloader %s not supported, available : %s" %
+                       (loader, ALL_LOADERS))
+    for module, loaders in supported_loaders.items():
+        if loader in loaders:
+            return module.create_loader(*args, **kwargs)
     raise RuntimeError("unexpected error! please report this as bug")
 
 
 # for maintainer, register your module here :
 # example:
 # ```
-# from .my_module import my_augmentation
-# register_module(my_augmentation)
+# from .my_module import my_loader
+# register_module(my_loader)
 # ```
-from . import albumentations
-register_module(module=albumentations)
-# register_module('albumentations') # still valid
+from . import pytorch_loader
+register_module(module=pytorch_loader)
+# register_module('pytorch_loader') # still valid
 
-from . import nvidia_dali
-register_module(module=nvidia_dali)
+from . import nvidia_dali_loader
+register_module(module=nvidia_dali_loader)
+
+